requirements.txt
nltk==3.5
pytorch
tqdm
matplotlib